+++ client/src/components/ProjectSamplesTable.js
@@ -10,8 +10,7 @@ import { useMyDataset } from 'hooks/useMyDataset'
 import { useProjectSamplesTable } from 'hooks/useProjectSamplesTable'
 import { differenceArray } from 'helpers/differenceArray'
 import { getProjectModalities } from 'helpers/getProjectModalities'
-import { getReadable } from 'helpers/getReadable'
-import { getReadableModality } from 'helpers/getReadableModality'
+import { getReadable, getReadableModality } from 'helpers/getReadable'
 import { DatasetAddSamplesModal } from 'components/DatasetAddSamplesModal'
 import { Icon } from 'components/Icon'
 import { Link } from 'components/Link'
@@ -204,7 +203,7 @@ export const ProjectSamplesTable = ({ stickies = 3, children }) => {
     if (!allSamples.length || !samples) return
 
     // Run only when the dataset contains data
-    if (!dataset.data && isDatasetDataEmpty) return
+    if (!dataset?.data && isDatasetDataEmpty) return
 
     // Use dataset on /datasets, otherwise use myDataset for setup
     const projectData = dataset
